# swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Bookstore API"
  description: "This is bookstore api documentation."
  contact:
    name: Bookstore Team
    url: https://github.com/nahidnstu12/bookmark-backend
    email: nahid.dev19@gmail.com
servers:
  - description: Dev
    url: "http://localhost:6050/api/v1"
...
paths:
...
  "/books":
    get:
      tags: [ Book ]
      description: Retrieve all published Books

      responses:
        "200":
          description: Return a list of books
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      allOf:
                        - type: object
                          properties:
                            id:
                              type: integer
                              format: int32
                              example: 1b9af82d-38cc-4191-8d9e-a1b9d7ca8f23
                            title:
                              type: string
                              example: Book name
                            publishedYear:
                              type: string
                              format: date
                              example: 2023-03-01 19:53:18
                            status:
                              type: number
                              format: number
                              example: 1
                            description:
                              type: string
                              format: text
                              example: New ABC 13 9370, 13.3, 5th Gen CoreA5-8250U, 8GB RAM, 256GB SSD,
                            bestSelling:
                              type: number
                              format: number
                              example: 1
                            totalPage:
                              type: number
                              format: number
                              example: 100
                            editions:
                              type: string
                              example: "[\"2011\"]"

                          required: [ id, title, description, publishedYear, status ]
                        - $ref: "#/components/schemas/Timestamp"

        "400":
          $ref: "#/components/responses/400"
        "500":
          "$ref": "#/components/responses/500"
    post:
      operationId: createBook
      security:
        - bearerAuth: [ ]
      tags: [ Book ]
      description: Create a new Book
      requestBody:
        description: To create a new book user must need to provide valid data
        content:
          "application/json":
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Book name
                publishedYear:
                  type: string
                  format: date
                  example: 2023-03-01 19:53:18
                status:
                  type: number
                  format: number
                  example: 1
                description:
                  type: string
                  format: text
                  example: New ABC 13 9370, 13.3, 5th Gen CoreA5-8250U, 8GB RAM, 256GB SSD,
                bestSelling:
                  type: number
                  format: number
                  example: 1
                totalPage:
                  type: number
                  format: number
                  example: 100
                editions:
                  type: string
                  format: json
                  example: "[\"2015\", \"2020\"]"
                
              required: [ id, title, description, publishedYear, status ]
      responses:
        "201":
          description: Successfully created a new book
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/BookCreateDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "500":
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Book:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int32
              example: 1b9af82d-38cc-4191-8d9e-a1b9d7ca8f23
            title:
              type: string
              example: Book name
            publishedYear:
              type: string
              format: date
              example: 2023-03-01 19:53:18
            status:
              type: number
              format: number
              example: 1
            description:
              type: string
              format: text
              example: New ABC 13 9370, 13.3, 5th Gen CoreA5-8250U, 8GB RAM, 256GB SSD,
            bestSelling:
              type: number
              format: number
              example: 1
            totalPage:
              type: number
              format: number
              example: 100
            editions:
              type: string
              example: "[\"2011\"]"
              
          required: [  name, description, published_year ]
        - $ref: "#/components/schemas/Timestamp"
    BookCreateDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 201
        message:
          type: string
          example: "Book created successfully"
        data:
          $ref: "#/components/schemas/Book"
...
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [createdAt, updatedAt]
...
  responses:
    "500":
      description: "Internal Server Error"
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              error:
                type: string
                example: Internal Server Error
              message:
                type: string
                example: We are sorry for the inconvenience. Please try again later
    "400":
      description: "Bad Request"
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              error:
                type: string
                example: Bad Request
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: field-name
                    message:
                      type: string
                      example: Invalid field-name
                    in:
                      type: string
                      enum: [query, body]
                      example: body
    "401":
      description: Unauthenticated
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthenticated
              message:
                type: string
                example: You need to login first
    "403":
      description: Unauthorized
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthorized
              message:
                type: string
                example: You don't have permission to do this
...